use anywidget for managing compatibility of widget with jupyter lab, jupyter notebook, and google colab
