+++ nbsv/viewer.py
@@ -7,8 +7,9 @@
 """
 TODO: Add module docstring
 """
-import ipywidgets
+import anywidget
 import numpy as np
+import pathlib
 from simulariumio import (
     TrajectoryData,
     TrajectoryConverter,
@@ -21,17 +22,11 @@ from simulariumio import (
 )
 from traitlets import Unicode, Int, Instance
 from typing import List, Dict
-from ._frontend import module_name, module_version
 
 
-@ipywidgets.register
-class ViewerWidget(ipywidgets.DOMWidget):
-    _model_name = Unicode("TrajectoryModel").tag(sync=True)
-    _model_module = Unicode(module_name).tag(sync=True)
-    _model_module_version = Unicode(module_version).tag(sync=True)
-    _view_name = Unicode("Viewport").tag(sync=True)
-    _view_module = Unicode(module_name).tag(sync=True)
-    _view_module_version = Unicode(module_version).tag(sync=True)
+class ViewerWidget(anywidget.AnyWidget):
+    _esm = pathlib.Path(__file__).parent / ".." / "lib" / "widget.js"
+    _css = pathlib.Path(__file__).parent / ".." / "lib" / "widget.css"
     trajectory = Instance(TrajectoryData)
     converter = Instance(TrajectoryConverter)
     trajectory_str = Unicode("").tag(sync=True)
@@ -54,7 +49,7 @@ class ViewerWidget(ipywidgets.DOMWidget):
             Height of the simularium viewer widget
             Default: 300
         """
-        super(ipywidgets.DOMWidget, self).__init__()
+        super().__init__()
         self.width = width
         self.height = height
         self.load_trajectory(trajectory)
