+++ package.json
@@ -5,7 +5,6 @@
   "keywords": [
     "jupyter",
     "jupyterlab",
-    "jupyterlab-extension",
     "widgets"
   ],
   "files": [
@@ -29,86 +28,24 @@
     "url": "https://github.com/simularium/nbsv"
   },
   "scripts": {
-    "build": "yarn run build:lib && yarn run build:nbextension && yarn run build:labextension:dev",
-    "build:prod": "yarn run build:lib && yarn run build:nbextension && yarn run build:labextension",
-    "build:labextension": "jupyter labextension build .",
-    "build:labextension:dev": "jupyter labextension build --development True .",
-    "build:lib": "tsc",
-    "build:nbextension": "webpack",
-    "clean": "yarn run clean:lib && yarn run clean:nbextension && yarn run clean:labextension",
-    "clean:lib": "rimraf lib",
-    "clean:labextension": "rimraf nbsv/labextension",
-    "clean:nbextension": "rimraf nbsv/nbextension/static/index.js",
-    "lint": "eslint . --ext .ts,.tsx --fix",
-    "lint:check": "eslint . --ext .ts,.tsx",
-    "prepack": "yarn run build:lib",
-    "test": "yarn jest --runInBand",
-    "watch": "npm-run-all -p watch:lib watch:nbextension watch:labextension",
-    "watch:lib": "tsc -w",
-    "watch:nbextension": "webpack --watch --mode=development",
-    "watch:labextension": "jupyter labextension watch ."
+    "dev": "npm run build -- --sourcemap=inline --watch",
+    "build": "bun build src/widget.tsx --minify --format=esm --outdir=./lib --asset-naming=[name].[ext]",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@aics/simularium-viewer": "^3.7.0",
-    "@ant-design/icons": "4.0.0",
-    "@jupyter-widgets/base": "^1.1.10 || ^2 || ^3 || ^4 || ^5 || ^6",
-    "antd": "^5.4.2",
-    "crypto": "1.0.1",
-    "jest-cli": "^29.7.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "@anywidget/react": "^0.0.8",
+    "@jupyter-widgets/base": "^6.0.10",
+    "antd": "^5.22.1",
+    "classnames": "^2.5.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "three": "^0.170.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.0",
-    "@babel/plugin-transform-runtime": "^7.21.4",
-    "@babel/preset-env": "^7.5.0",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.4",
-    "@jest/globals": "^29.5.0",
-    "@jupyter-widgets/base-manager": "^1.0.2",
-    "@jupyterlab/builder": "^3.6.4",
-    "@lumino/application": "^2.1.0",
-    "@lumino/domutils": "^2.0.0",
-    "@lumino/widgets": "^2.1.0",
-    "@types/jest": "^29.5.0",
-    "@types/react": "^18.0.34",
-    "@types/react-dom": "^18.0.11",
-    "@types/webpack-env": "^1.13.6",
-    "@typescript-eslint/eslint-plugin": "^6",
-    "@typescript-eslint/parser": "^6",
-    "acorn": "^7.2.0",
-    "babel-loader": "^9.1.2",
-    "babel-plugin-transform-import-meta": "^2.2.0",
-    "browser-env": "^3.3.0",
-    "css-loader": "^3.2.0",
-    "eslint": "^7.4.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "fs-extra": "^7.0.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jsdom-global": "^3.0.2",
-    "mkdirp": "^0.5.1",
-    "npm-run-all": "^4.1.3",
-    "prettier": "^2.0.5",
-    "rimraf": "^2.6.2",
-    "source-map-loader": "^1.1.3",
-    "style-loader": "^1.0.0",
-    "ts-jest": "^29.1.0",
-    "ts-loader": "^8.0.0",
-    "typescript": "^5.0.4",
-    "webpack": "^5.61.0",
-    "webpack-cli": "^5.1.3"
-  },
-  "jupyterlab": {
-    "extension": "lib/plugin",
-    "outputDir": "nbsv/labextension/",
-    "sharedPackages": {
-      "@jupyter-widgets/base": {
-        "bundled": false,
-        "singleton": true
-      }
-    }
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.1",
+    "@types/three": "^0.169.0",
+    "typescript": "^5.6.3",
+    "@types/bun": "latest"
   }
 }
