+++ tsconfig.json
@@ -1,26 +1,25 @@
 {
-  "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop":true,
-    "jsx": "react",
-    "lib": ["es2017", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noUnusedLocals": true,
-    "outDir": "lib",
-    "resolveJsonModule": true,
-    "rootDir": "src",
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "strict": true,
-    "strictPropertyInitialization": false,
-    "target": "es2015",
-    "types": ["jest"]
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx", "browser-env.d.ts",
-  ],
-  "exclude": ["src/**/__tests__"]
+	"include": [
+		"src"
+	],
+	"compilerOptions": {
+		"target": "ES2020",
+		"module": "ESNext",
+		"lib": [
+			"ES2020",
+			"DOM",
+			"DOM.Iterable"
+		],
+		"skipLibCheck": true,
+		"moduleResolution": "bundler",
+		"allowImportingTsExtensions": true,
+		"resolveJsonModule": true,
+		"isolatedModules": true,
+		"noEmit": true,
+		"jsx": "react",
+		"strict": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noFallthroughCasesInSwitch": true
+	}
 }
